Enable `full` notebook windowing mode by default (#15964)
* Enable `full` windowing mode by default

* Use `none` mode for mermaid snapshots tests

* Account for the windowed scrollbar button

* Do not capture the full main area in galata notebook tests

it does not make much sense, especially for cells; instead,
capture the notebook itself. Add a new test for saving,
because that was implicitly tested by `Run Cells` test
(and is not anymore). Of note, one of the tests,
`Open and run cell by cell` is still taking a snapshot
of the entire panel; this is intentional - we do want
one snapshot to be more encompassing (but not all of them!).

* Update the debugger button sizing

* Only capture notebook content in custom CSS test of headings

as the toolbar is irrelevant for the snapshot of headings!

* Only capture the main area for altair and html docs snapshots

The rest of the UI does not really matter on these snapshots
and it only makes the content small and hard to read in the docs.

* Update Playwright Snapshots

* Update Playwright Snapshots

* Fix toolbar test (account for the scrollbar button)

* Revert spurious snapshot updates

* Fix missing dot in selector

* Use `none` for collapsible headings test

* Add notebook-tab-saved snapshot

* Fix misplaced config

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>