Fix `docmanager/savehandler` "continue to save" test (#16933)
There were two problems with the previous test so that it didn't really
test what was intended.

When the signal was triggered and the test() method was called, the
promise was resolved regardless of test()'s return value. The subsequent
change to the context.model inside test() then did not trigger anything.  We
switch to find() which only resolves the promise when the return value is true.

Additionally, the initial call to the triggered function would always return
true due to the ordering of the increment and check of the call counter; this
has also been fixed.