Fix total size estimation in full windowing mode to reduce scrollbar jitter (#16950)
* Fix windowing total size estimation to reduce scrollbar jitter

1) Previously if height of a cell in the middle of the notebook changed
then the sizes and offsets of all subsequent cells were discarded.
After change the offset delta is instead computed and subsequent sizes
and offsets are adjusted rather than being discarded.

2) Previously the size estimates were only stored after the viewport
window was moved away from initial null position; this constraint was
now reduced to also store sizes when windowing is active but window
size has not been computed yet.

* Fix the case of adding/removing cells and heal the offsets

after changes to the list. Also, rename `lastMeasuredIndex`
to `measuredAllUntilIndex` to better reflect that we are
storing measurements for cells that are after this point
and we can use those measurements (just cannot use
the faster algorithm for finding the index from offset).

* Stylistic improvements to comments and tests

* Apply suggestions from code review

Co-authored-by: Jason Weill <93281816+JasonWeill@users.noreply.github.com>

* `stopSize` → `stopSizer`

---------

Co-authored-by: Jason Weill <93281816+JasonWeill@users.noreply.github.com>